+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog for 87th Admin Bot
 Note: This file was started on 2023-04-22. Any changes before then are not captured.
 
+## 2023-12-15
+- Updated event announcement text
+
 ## 2023-12-10
 - Updated README
 - Renamed `enums.py` to `configs.py` to better represent content of file